requirements.txt
asgiref==3.4.1
certifi==2021.10.8
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
djangorestframework==3.12.4
django-cors-headers==3.11.0
gunicorn==20.1.0
jsonfield==3.1.0
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
